feat(api): Added confirmation email whenever a User become a host
