+++ apps/api/src/routes/users/service/authEmail.ts
@@ -2,7 +2,7 @@ import { render } from '@react-email/render'
 import { EmailService } from '@/common/service/email'
 import ForgotPasswordEmail from '@/routes/users/email-template/auth/ForgotPasswordEmail'
 import MultiFactorAuth from '@/routes/users/email-template/auth/MultiFactorAuth'
-import HostConfirmationEmail from '@/routes/users/email-template/host/HostConfirmationEmail'
+import BecomeHostEmailConfirmation from '@/routes/users/email-template/host/BecomeHostEmailConfirmation'
 
 export type TSendEmailParams = {
   to: string[] | string
@@ -28,17 +28,6 @@ export class AuthEmail extends EmailService {
       to: typeof to === 'string' ? [to] : to,
       template: emailHtml,
       subject: 'Multi-factor Authentication Code',
-    })
-    return await sendEmail
-  }
-
-  async sendHostConfirmation(sendEmailParams: TSendEmailParams) {
-    const { to } = sendEmailParams
-    const emailHtml = render(HostConfirmationEmail())
-    const sendEmail = super.sendEmail({
-      to: typeof to === 'string' ? [to] : to,
-      template: emailHtml,
-      subject: 'ExploreSiargao Hosting',
     })
     return await sendEmail
   }
